Safari-611.1.10/LayoutTests/editing/deleting/ios/backspace-last-character-expected.txt

CONSOLE MESSAGE: Activating element and waiting for input session.
CONSOLE MESSAGE: Waiting for element focus.
CONSOLE MESSAGE: Typing 'w'.
CONSOLE MESSAGE: Waiting for first input event.
CONSOLE MESSAGE: Hitting backspace.
CONSOLE MESSAGE: Waiting for second input event.
CONSOLE MESSAGE: Test complete.
This tests using the software keyboard and pressing 'w' then backspace deletes the 'w' and does not cause an assertion failure in a debug build.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS inputDidFocus became true
PASS inputEventCount became 1
PASS document.getElementById("input").value is "w"
PASS inputEventCount became 2
PASS document.getElementById("input").value is ""
PASS successfullyParsed is true

TEST COMPLETE
